\name{LPIparam.default.gdf}

\title{LPIparam.default.gdf}

\alias{LPIparam.default.gdf}

\description{LPI default parameters}

\usage{LPIparam.default.gdf()}

\arguments{None}

\value{
  A named list of LPI.gdf default arguments. See \link{LPI.gdf} for details.
}

\author{Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}}
