\name{LPIsaveACF.gdf}
\title{LPIsaveACF.gdf}
\alias{LPIsaveACF.gdf}
\description{ Save LPI results to files}
\usage{LPIsaveACF.gdf( LPIparam , intPeriod , ACF ) }

\arguments{
  \item{ LPIparam }{ An LPI parameter list from \link{LPI.gdf}}
  \item{ intPeriod}{ Integration period number.}
  \item{ ACF }{ An ACF list. See LPIsaveACF from the main LPI package
    for details.}
}

\value{
  Result file name (invisibly)
}




  \author{Ilkka Virtanen (University of Oulu, Finland) \cr
  \email{ilkka.i.virtanen@oulu.fi}}
